wip - removing test_component.py, adding tests and todos for testing as_dict implementation
